# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SED Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-units:
    get:
      operationId: listBusinessUnits
      tags: [business-units]
      summary: List all business units
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
            enum: [Media, Digital, Tourism, Luxury, "Travel Support", Lifestyle, Service, Retail, "null"]
      responses:
        "200":
          description: List of business units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessUnit"
    post:
      operationId: createBusinessUnit
      tags: [business-units]
      summary: Create a business unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessUnit"
  /business-units/{id}:
    get:
      operationId: getBusinessUnit
      tags: [business-units]
      summary: Get business unit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business unit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessUnitDetail"
        "404":
          description: Not found
    patch:
      operationId: updateBusinessUnit
      tags: [business-units]
      summary: Update business unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessUnit"


  # ── CUSTOMERS (CRM) ────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [Lead, Prospect, Negotiation, Customer, "Repeat Customer", VIP, "null"]
        - name: businessUnitId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    BusinessUnit:
      type: object
      required: [id, name, category, slug, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        category:
          type: string
          enum: [Media, Digital, Tourism, Luxury, "Travel Support", Lifestyle, Service, Retail]
        website: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    BusinessUnitDetail:
      allOf:
        - $ref: "#/components/schemas/BusinessUnit"
        - type: object
          properties:
            totalRevenue: { type: number }
            totalExpenses: { type: number }
            netProfit: { type: number }
            totalCustomers: { type: integer }
            totalProjects: { type: integer }
            totalInvoices: { type: integer }
            totalStaff: { type: integer }
            notes: { type: ["string", "null"] }
...
    BusinessUnitInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        slug: { type: string }
        category:
          type: string
          enum: [Media, Digital, Tourism, Luxury, "Travel Support", Lifestyle, Service, Retail]
        website: { type: string }
        logoUrl: { type: string }
        description: { type: string }
        isActive: { type: boolean }
        notes: { type: string }
    BusinessUnitUpdate:
      type: object
      properties:
        name: { type: string }
        category:
          type: string
          enum: [Media, Digital, Tourism, Luxury, "Travel Support", Lifestyle, Service, Retail]
        website: { type: string }
        logoUrl: { type: string }
        description: { type: string }
        isActive: { type: boolean }
        notes: { type: string }

    # ── CUSTOMERS ─────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, customerId, fullName, status, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: string }
        fullName: { type: string }
        businessName: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        nationality: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        businessUnitId: { type: ["integer", "null"] }
        businessUnitName: { type: ["string", "null"] }
        serviceInterested: { type: ["string", "null"] }
        leadSource: { type: ["string", "null"] }
        facebook: { type: ["string", "null"] }
        instagram: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status:
          type: string
          enum: [Lead, Prospect, Negotiation, Customer, "Repeat Customer", VIP]
        lastContact: { type: ["string", "null"], format: date }
        nextFollowUp: { type: ["string", "null"], format: date }
        assignedStaff: { type: ["string", "null"] }
        tags: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CustomerList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    CustomerInput:
      type: object
      required: [fullName, status]
      properties:
        fullName: { type: string }
        businessName: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        email: { type: string }
        nationality: { type: string }
        country: { type: string }
        address: { type: string }
        businessUnitId: { type: integer }
        serviceInterested: { type: string }
        leadSource: { type: string }
        facebook: { type: string }
        instagram: { type: string }
        website: { type: string }
        notes: { type: string }
        status:
          type: string
          enum: [Lead, Prospect, Negotiation, Customer, "Repeat Customer", VIP]
        lastContact: { type: string, format: date }
        nextFollowUp: { type: string, format: date }
        assignedStaff: { type: string }
        tags: { type: string }
    CustomerUpdate:
      type: object
      properties:
        fullName: { type: string }
        businessName: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        email: { type: string }
        nationality: { type: string }
        country: { type: string }
        address: { type: string }
        businessUnitId: { type: integer }
        serviceInterested: { type: string }
        leadSource: { type: string }
        facebook: { type: string }
        instagram: { type: string }
        website: { type: string }
        notes: { type: string }
        status:
          type: string
          enum: [Lead, Prospect, Negotiation, Customer, "Repeat Customer", VIP]
        lastContact: { type: string, format: date }
        nextFollowUp: { type: string, format: date }
        assignedStaff: { type: string }
        tags: { type: string }

    # ── SALES / PIPELINE ──────────────────────────────────────────────────────
...