# lib/api-zod/src/generated/types/customerList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SED Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';

export interface CustomerList {
  data: Customer[];
  total: number;
  page: number;
  limit: number;
}