# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SED Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateStatus } from './customerUpdateStatus';

export interface CustomerUpdate {
  fullName?: string;
  businessName?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  nationality?: string;
  country?: string;
  address?: string;
  businessUnitId?: number;
  serviceInterested?: string;
  leadSource?: string;
  facebook?: string;
  instagram?: string;
  website?: string;
  notes?: string;
  status?: CustomerUpdateStatus;
  lastContact?: Date;
  nextFollowUp?: Date;
  assignedStaff?: string;
  tags?: string;
}